pyproject.toml
[project]
name = "bl-odoo"
version = "0.1.1"
description = "A command-line tool for managing Odoo dependencies."
authors = [
  { name="Your Name", email="your.email@example.com" },
]
dependencies = [
    "pyyaml>=6.0.3",
    "rich",
    "typer",
]
requires-python = ">=3.9"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]

[project.optional-dependencies]
dev = [
    "ruff",
    "kalong",
]

[project.scripts]
bl = "bl.__main__:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bl"]

[tool.setuptools.package-data]
"bl" = ["py.typed"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
